Filter data frame for relevant results

The filter_select_name() function filters a data frame for relevant results, based on the scientific name given. Filtering can be done with scripts by exact or fuzzy match. Or, for a more controlled approach, this function provides interactive filtering by providing the user with prompts. The interactive method allows the user to manually determine whether they wish to keep results containing certain scientific names.

Usage

filter_select_name(df, synonyms_list, filter, accepted_name)

Arguments

df

Data frame with name column to be fixed.

synonyms_list

A list of synonyms for a species.

filter

The type of filter to be used--either "exact", "fuzzy", or "interactive".

accepted_name

The accepted scientific name for the species.

Value

Returns data frame with filtered results.

Details

This function requires packages dplyr and magrittr.