Suppress print statements and messages

The suppress_output() function suppresses print statements and messages from a function. This does not suppress warnings or errors.

Usage

suppress_output(input)

Arguments

input

Value

output Output is the output of the passed function call.

Details

This function requires no additional packages.