Cite Data - Get species, longitude, and latitude columns

The citation_bellow function retrieves and returns the citation information for the data provided by GBIF in a data frame.

Usage

citation_bellow(df, id = "ID", aggregator = "aggregator")

Arguments

df

Data frame of occurrence records returned from gators_download().

id

Default = "ID". The name of the id column in the data frame, which contains unique IDs defined from GBIF or iDigBio.

aggregator

Default = "aggregator". The name of the column in the data frame that identifies the aggregator that provided the record.

Value

Returns a list with citation information for the GBIF data downloaded.

Details

This function requires the rgbif package.

Examples

data <- data[1800:1900, ]
citations <- citation_bellow(data)
#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5
#> [1] 6
#> [1] 7
#> [1] 8
#> [1] 9
#> [1] 10
#> [1] 11
#> [1] 12
#> [1] 13
#> [1] 14
#> [1] 15
#> [1] 16
#> [1] 17
#> [1] 18
#> [1] 19
#> [1] 20
#> [1] 21
#> [1] 22
#> [1] 23
#> [1] 24
#> [1] 25
#> [1] 26
#> [1] 27
#> [1] 28
#> [1] 29
#> [1] 30
#> [1] 31
#> [1] 32
#> [1] 33
#> [1] 34
#> [1] 35
#> [1] 36
#> [1] 37
#> [1] 38
#> [1] 39
#> [1] 40
#> [1] 41
#> [1] 42
#> [1] 43
#> [1] 44
#> [1] 45
#> [1] 46
#> [1] 47
#> [1] 48
#> [1] 49
#> [1] 50
#> [1] 51
#> [1] 52
#> [1] 53
#> [1] 54
#> [1] 55
#> [1] 56
#> [1] 57
#> [1] 58
#> [1] 59
#> [1] 60
#> [1] 61
#> [1] 62
#> [1] 63
#> [1] 64
#> [1] 65
#> [1] 66
#> [1] 67
#> [1] 68