Download specimen data from both iDigBio and GBIF

The gators_download() function downloads data from GBIF and iDigBio for your desired species. Data returned includes: scientificName, genus, specificEpithet, infraspecificEpithet, basisOfRecord, eventDate, institutionCode, collectionCode, collectionID, country, county, stateProvince, locality, latitude, longitude, identificationID, coordinateUncertaintyInMeters, informationWithheld, and habitat.

Usage

gators_download(
  synonyms_list,
  newFileName,
  gbif_match = "fuzzy",
  idigbio_filter = TRUE
)

Arguments

synonyms_list

A list of synonyms for your desired species. For example, synonyms_list = c("Asclepias curtissii","Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii"). This parameter is required.

newFileName

The path and file name for the retrieved data. Note that this parameter should include the ".csv" extension as well. For example, newFileName = "base_folder/other_folder/my_file.csv". The file path can be entered either as relative to the current working directory (example: "../my_file.csv") or as a full path. This parameter is required.

gbif_match

A parameter to select either search by fuzzy matching of scientific name or to search by species code. For example, gbif_match = "fuzzy" will search by fuzzy match and gbif_match = "code" will search by code. This parameter is not required and is assigned "fuzzy" by default.

idigbio_filter

A parameter to remove less relevant search results from iDigBio. Based on the search input, results may include data points for a different species that mention the desired species in the locality information, for example. Choosing idigbio_filter = TRUE will return the data frame with rows in which the name column fuzzy matches a name on the synonym list. This parameter is not required and is assigned TRUE by default.

Value

Writes a csv file as specified in the input. This csv file will contain search results for the desired species from the GBIF and iDigBio databases. The columns are as follows:

scientificName

genus

specificEpithet

infraspecificEpithet

basisOfRecord

eventDate

institutionCode

collectionCode

collectionID

identificationID

informationWithheld

country

county

stateProvince

locality

latitude

longitude

coordinateUncertaintyInMeters

habitat

Details

This function uses the getidigbio(), getgbif(), fix_columns(), fix_names(), and filter_fix_names() functions. This function requires packages magrittr, rgbif, dplyr, ridigbio, and stringr.

Examples

gators_download(c("Asclepias curtissii", "Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii"), "my_new_file.csv", idigbio_filter = FALSE)
#> [1] "Warning: iDigBio search will return all records where any column has a matching string to the provided scientific names."
gators_download(c("Asclepias curtissii", "Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii"), "newFile.csv", gbif_match = "code")